fix(cloudformation): render variables in cfn vertices config (#7423)
* render cfn vertices config varaibles

* mypy

* export common function

* export function

* lint

* move unit tests

* rename file
